Simplified logic ...
'if [ z ...' kind of like 'if ! ...' but harder to see quickly.  By setting
the variable to 'true' or 'false', I can use them in a way that is more
expressive.  Opportunities to simplif the logic become easier to spot:

if !A && !B ; X ; else Y can become if A || B ; Y else X